some update

+++ server/api/admin/menu/[uuid].delete.ts
@@ -1,14 +1,18 @@
 import prisma from "~/prisma/client";
 import { MENU_DELETED } from "~/server/message";
+import { jwtCheckerHelper } from "~/server/helper/jwt-checker.helper";
 
 export default defineEventHandler(async (event) => {
+    let user = await jwtCheckerHelper(event);
     const uuid = event.context.params?.uuid;
 
-    await prisma.menu.delete({
-        where: {
-            uuid: uuid
-        }
-    });
+    if (user) {
+        await prisma.menu.delete({
+            where: {
+                uuid: uuid
+            }
+        });
 
-    return MENU_DELETED;
+        return MENU_DELETED;
+    }
 });
